requirements.txt
python-dotenv>=1.0
openai>=1.40.0
langchain>=0.2.14
langchain-openai>=0.1.22
qdrant-client>=1.9.0
pypdf>=4.2.0
tiktoken>=0.7.0
streamlit>=1.35.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
pandas>=2.2.0
